+++ packages/www/components/DocsLayout/index.tsx
@@ -1,4 +1,4 @@
-import Layout from "../Layout";
+import Layout from "../../layouts";
 import { Flex, Box, Container } from "@theme-ui/components";
 import { useState } from "react";
 import TableOfContents from "../TableOfContents";
