+++ packages/www/components/TabbedLayout/index.tsx
@@ -1,4 +1,4 @@
-import Layout from "../../components/Layout";
+import Layout from "../../layouts";
 import { Flex } from "@theme-ui/components";
 import Tabs, { TabType } from "../../components/Tabs";
 import { FunctionComponent } from "react";
