+++ packages/www/pages/app/stream/[id].tsx
@@ -13,7 +13,7 @@ import {
   Alert,
   Close,
 } from "@theme-ui/components";
-import Layout from "../../../components/Layout";
+import Layout from "../../../layouts";
 import useLoggedIn from "../../../hooks/use-logged-in";
 import { Stream } from "@livepeer.com/api";
 import { CopyToClipboard } from "react-copy-to-clipboard";
