+++ packages/www/pages/docs/index.tsx
@@ -1,5 +1,5 @@
 import Fade from "react-reveal/Fade";
-import Layout from "../../components/Layout";
+import Layout from "../../layouts";
 import { Container } from "@theme-ui/components";
 import DocsCategoryCard from "../../components/cards/cards/category";
 import KeenSliderGrid from "../../components/KeenSliderGrid";
