+++ packages/www/pages/jobs/[slug].tsx
@@ -4,7 +4,7 @@ import { print } from "graphql/language/printer";
 import ReactMarkdown from "react-markdown";
 import Fade from "react-reveal/Fade";
 import Button from "../../components/Button";
-import Layout from "../../components/Layout";
+import Layout from "../../layouts";
 import Prefooter from "../../components/Prefooter";
 import allJobs from "../../queries/allJobs.gql";
 import Code from "../../components/renderers/Code";
