+++ packages/www/pages/usage.tsx
@@ -1,5 +1,5 @@
 import { Box, Heading, Container } from "@theme-ui/components";
-import Layout from "../components/Layout";
+import Layout from "../layouts";
 
 const Usage = () => {
   return (
